requirements.txt
# ==========================================
# CORE DJANGO STACK
# ==========================================
Django==5.1.4
psycopg2-binary==2.9.10
django-environ==0.11.2
django-extensions==3.2.3
dj-database-url==2.3.0
whitenoise==6.8.2

# ==========================================
# REST API
# ==========================================
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
drf-yasg==1.21.11

# ==========================================
# AI & ML STACK 
# ==========================================
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.8.0+cpu
openai==1.58.1
sentence-transformers==3.3.1
pgvector==0.3.6
transformers==4.56.1

# ==========================================
# DATA PROCESSING
# ==========================================
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1

# ==========================================
# EXTERNAL APIS
# ==========================================
requests==2.32.3
tmdbv3api==1.9.0
python-dotenv==1.0.1
httpx==0.28.1

# ==========================================
# WEB DEVELOPMENT
# ==========================================
gunicorn==23.0.0
pillow==11.0.0

# ==========================================
# UTILITIES
# ==========================================
python-dateutil==2.9.0.post0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
python-decouple==3.8